Refactor Delayed Tasks to their own class
