+++ fml/BUILD.gn
@@ -17,6 +17,8 @@ source_set("fml") {
     "concurrent_message_loop.cc",
     "concurrent_message_loop.h",
     "eintr_wrapper.h",
+    "delayed_task.cc",
+    "delayed_task.h",
     "file.cc",
     "file.h",
     "icu_util.cc",
